- refs #81966: Renamed map component for the download section. - refs #81966: Added initial version of the RasterDownloadClimateVariable class.
